config: refactor, extract configSchema
