+++ packages/config/index.ts
@@ -1,3 +1,5 @@
 // Path: packages/config/index.ts
 
 export * from './src/config';
+export * from './src/configSchema';
+export * from './src/merge';
